requirements.txt
requests==2.20.0
boto3==1.9.243
google-cloud-storage==1.20.0